Cleaned up the LZ77 code and remove the TokenReader and TokenWriter classes.

+++ code/lz77/lz77_dec.h
@@ -3,6 +3,10 @@
 
 #include <cstdio>
 
-void decompress(signed long windowSize, signed long lookAheadSize, FILE * in, FILE * out);
+void decompress(
+    unsigned int windowSize,
+    unsigned int lookAheadSize,
+    FILE * in,
+    FILE * out);
 
 #endif /* _LZ77_DEC_H_ */
